requirements.txt
fastapi
uvicorn
openai
python-dotenv
sentence-transformers
pinecone
langchain
bs4
requests
chromadb